Clearly define start & end of match with UI/state
In order to fix a problem with rendering the graph (discussed below), a
match startTime and endTime property was added to the matchDetails state,
which were set using new "start game" and "match complete" buttons. match
duration and date were also added for future use in filtering matches
inside the Dashboard.

There was a problem whereby the graph wouldn't be able to handle either
one set of heart rate date or none at all (it could only handle both
sets). This was because the baseTime was defined using both datasets,
this has now been changed to utilise the match starting time. The lastTime
variable worked in a similar way, it is now defined differently depending
on the number of datasets available.