Associate relevant fetch requests to match start
- Utilised refs to store playerIds and matchId for future API calls
- Linked creation of player records to "Start Game" button on input screen
- Linked creation of match and match_player records to "Start Match"
  button inside Game Tracking component

Note: This took longer than expected due to a problem with integer sizes
being stored inside of the database. It turns out storing the date in
milliseconds was too large an integer for the regular integer datatype,
causing an error when trying to create match records. Rather than
changing to a BIG INT datatype, the date is stored in seconds instead.